+++ cirq/protocols/__init__.py
@@ -46,7 +46,7 @@ from cirq.protocols.decompose_protocol import (
     SupportsDecompose,
     SupportsDecomposeWithQubits,
 )
-from cirq.protocols.equal_up_to_global_phase import (
+from cirq.protocols.equal_up_to_global_phase_protocol import (
     equal_up_to_global_phase,)
 from cirq.protocols.inverse_protocol import (
     inverse,)
@@ -59,7 +59,7 @@ from cirq.protocols.measurement_key_protocol import (
     is_measurement,
     measurement_key,
 )
-from cirq.protocols.mixture import (
+from cirq.protocols.mixture_protocol import (
     has_mixture,
     has_mixture_channel,
     mixture,
@@ -70,7 +70,7 @@ from cirq.protocols.mixture import (
 from cirq.protocols.mul import (
     mul,
 )
-from cirq.protocols.pauli_expansion import (
+from cirq.protocols.pauli_expansion_protocol import (
     pauli_expansion,
 )
 # pylint: disable=redefined-builtin
@@ -109,7 +109,7 @@ from cirq.protocols.unitary import (
     SupportsUnitary,
     unitary,
 )
-from cirq.protocols.has_unitary import (
+from cirq.protocols.has_unitary_protocol import (
     has_unitary,
     SupportsExplicitHasUnitary,
 )
