+++ cirq/protocols/channel.py
@@ -20,7 +20,7 @@ from typing import Any, Sequence, Tuple, TypeVar, Union
 import numpy as np
 from typing_extensions import Protocol
 
-from cirq.protocols.mixture import has_mixture_channel
+from cirq.protocols.mixture_protocol import has_mixture_channel
 
 
 from cirq.type_workarounds import NotImplementedType
