+++ cirq/protocols/equal_up_to_global_phase_protocol.py
