+++ cirq/protocols/has_unitary_protocol.py
