+++ cirq/protocols/has_unitary_protocol_test.py
