+++ cirq/protocols/mixture_protocol_test.py
