+++ cirq/protocols/pauli_expansion_protocol.py
