+++ cirq/protocols/pauli_expansion_protocol_test.py
